+++ changelog.md
@@ -2,6 +2,8 @@
 
 ## Version 4.0.3
  - adding css class to input if no value is selected and text is inserted in input
+ - support for using free text as value of typeahead
+ - support for specifying displayedProperty using `string` or `function`
 
 ## Version 4.0.2
  - returned typescript version back to 2.4.2 and removed distJit